refactor: Expose mountpoints types in OCP
Avoid having some special checks in OC about classes defined in OCA.
